+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Added stat cards to organization view
 - Added a clear icon to search input and removed dead student performance view
 - Added a message to search results when no results are found